requirements.txt
pandas
torch
torchvision
numpy
timm
huggingface_hub
tqdm
Pillow
pyarrow
